make necesssary changes to controller service repository

+++ src/main/java/com/dev/soulScript_Api/service/impl/JournalServiceImpl.java
@@ -1,8 +1,17 @@
 package com.dev.soulScript_Api.service.impl;
 
+import com.dev.soulScript_Api.dto.JournalEntryRequestDTO;
+import com.dev.soulScript_Api.dto.JournalEntryResponseDTO;
+import com.dev.soulScript_Api.exception.JournalEntryNotFoundException;
+import com.dev.soulScript_Api.exception.UserNotFoundException;
+import com.dev.soulScript_Api.mapper.JournalEntryMapper;
 import com.dev.soulScript_Api.model.JournalEntry;
+import com.dev.soulScript_Api.model.User;
 import com.dev.soulScript_Api.repository.JournalRepository;
+import com.dev.soulScript_Api.repository.UserRepository;
 import com.dev.soulScript_Api.service.JournalService;
+import jakarta.validation.Valid;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -10,30 +19,45 @@ import java.util.Optional;
 
 @Service
 public class JournalServiceImpl implements JournalService {
+
     private final JournalRepository journalRepository;
+    private final UserRepository userRepository;
 
-
-    public JournalServiceImpl(JournalRepository journalRepository) {
+    @Autowired
+    public JournalServiceImpl(JournalRepository journalRepository, UserRepository userRepository) {
         this.journalRepository = journalRepository;
+        this.userRepository = userRepository;
     }
 
-    public List<JournalEntry> getAllEntries() {
-        return journalRepository.findAll();
+    public List<JournalEntryResponseDTO> getAllEntries() {
+        List<JournalEntry> entries = journalRepository.findAll();
+        return entries.stream().map((entry) -> JournalEntryMapper.toResponse(entry)).toList();
     }
 
-    public List<JournalEntry> getEntriesByUser(Long userId) {
-        return journalRepository.findByUserId(userId);
+    public List<JournalEntryResponseDTO> getEntriesByUser(String username){
+        List<JournalEntry> entries = journalRepository.findByUserUsername(username);
+        return entries.stream().map((entry) -> JournalEntryMapper.toResponse(entry)).toList();
     }
 
-    public Optional<JournalEntry> getEntryById(Long id) {
-        return journalRepository.findById(id);
+    public JournalEntryResponseDTO getEntryById(Long id) {
+        Optional<JournalEntry> entry = journalRepository.findById(id);
+        if(entry.isEmpty()) throw new JournalEntryNotFoundException("Journal Entry for id "+id+" not found.");
+
+        return JournalEntryMapper.toResponse(entry.get());
     }
 
-    public JournalEntry saveEntry(JournalEntry entry) {
-        return journalRepository.save(entry);
+    public JournalEntryResponseDTO saveEntry(JournalEntryRequestDTO request, String username){
+        Optional<User> user = userRepository.findByUsername(username);
+        if(user.isEmpty()) throw new UserNotFoundException("User not found for given username");
+
+        JournalEntry journal = JournalEntryMapper.toEntity(request,user.get());
+        return JournalEntryMapper.toResponse(journalRepository.save(journal));
     }
 
     public void deleteEntry(Long id) {
+        if(!journalRepository.existsById(id)){
+            throw new JournalEntryNotFoundException("Journal for id "+id+" not found");
+        }
         journalRepository.deleteById(id);
     }
 }
